Make PID an int instead of a byte for longer commands; add mode override for additional modes (e.g. 22 for Chevy Volt)

+++ OBD.NET/OBD.NET.Common/OBDData/IOBDData.cs
@@ -2,7 +2,7 @@
 {
     public interface IOBDData
     {
-        byte PID { get; }
+        int PID { get; }
 
         void Load(string data);
     }
